# optimize `is_ipv4`