## Before

```
BasicBench_AdaURL_url_aggregator     173992 ns       173979 ns         4074    3.24877     27.344    660.621           110.716              4.049         13.1542         2.67485k    203.345 118.313M/s  8.45217ns  204.201ns 4.89713M/s
```

## After

```
BasicBench_AdaURL_url_aggregator     168135 ns       168133 ns         4156     3.2483    26.3602    636.853           109.969            4.17178         13.5512         2.65681k    196.058 122.427M/s  8.16814ns  197.339ns 5.06742M/s
```